+++ src/conversion/to_geo_types.rs
@@ -18,7 +18,7 @@ where
     fn try_from(value: &geometry::Value) -> Result<Self> {
         match value {
             geometry::Value::Point(point_type) => Ok(create_geo_point(&point_type)),
-            _ => Err(Error::InvalidGeometryConversion(value.clone())),
+            other => Err(mismatch_geom_err("Point", &other)),
         }
     }
 }
@@ -39,7 +39,7 @@ where
                     .map(|point_type| create_geo_point(&point_type))
                     .collect(),
             )),
-            _ => Err(Error::InvalidGeometryConversion(value.clone())),
+            other => Err(mismatch_geom_err("MultiPoint", &other)),
         }
     }
 }
@@ -57,7 +57,7 @@ where
             geometry::Value::LineString(multi_point_type) => {
                 Ok(create_geo_line_string(&multi_point_type))
             }
-            _ => Err(Error::InvalidGeometryConversion(value.clone())),
+            other => Err(mismatch_geom_err("LineString", &other)),
         }
     }
 }
@@ -75,7 +75,7 @@ where
             geometry::Value::MultiLineString(multi_line_string_type) => {
                 Ok(create_geo_multi_line_string(&multi_line_string_type))
             }
-            _ => Err(Error::InvalidGeometryConversion(value.clone())),
+            other => Err(mismatch_geom_err("MultiLineString", &other)),
         }
     }
 }
@@ -91,7 +91,7 @@ where
     fn try_from(value: &geometry::Value) -> Result<Self> {
         match value {
             geometry::Value::Polygon(polygon_type) => Ok(create_geo_polygon(&polygon_type)),
-            _ => Err(Error::InvalidGeometryConversion(value.clone())),
+            other => Err(mismatch_geom_err("Polygon", &other)),
         }
     }
 }
@@ -109,12 +109,11 @@ where
             geometry::Value::MultiPolygon(multi_polygon_type) => {
                 Ok(create_geo_multi_polygon(&multi_polygon_type))
             }
-            _ => Err(Error::InvalidGeometryConversion(value.clone())),
+            other => Err(mismatch_geom_err("MultiPolygon", other)),
         }
     }
 }
 try_from_owned_value!(geo_types::MultiPolygon<T>);
-
 
 #[cfg_attr(docsrs, doc(cfg(feature = "geo-types")))]
 impl<T> TryFrom<&geometry::Value> for geo_types::GeometryCollection<T>
@@ -133,7 +132,7 @@ where
 
                 Ok(geo_types::GeometryCollection(geojson_geometries))
             }
-            _ => Err(Error::InvalidGeometryConversion(value.clone())),
+            other => Err(mismatch_geom_err("GeometryCollection", other)),
         }
     }
 }
@@ -351,6 +350,13 @@ where
             .map(|polygon_type| create_geo_polygon(&polygon_type))
             .collect(),
     )
+}
+
+fn mismatch_geom_err(expected_type: &'static str, found: &geometry::Value) -> Error {
+    Error::InvalidGeometryConversion {
+        expected_type,
+        found_type: found.type_name(),
+    }
 }
 
 #[cfg(test)]
@@ -609,6 +615,19 @@ mod tests {
     }
 
     #[test]
+    fn geojson_mismatch_geometry_conversion_test() {
+        let coord1 = vec![100.0, 0.2];
+        let coord2 = vec![101.0, 1.0];
+        let geojson_line_string = Value::LineString(vec![coord1.clone(), coord2.clone()]);
+        use std::convert::TryFrom;
+        let error = geo_types::Point::<f64>::try_from(geojson_line_string).unwrap_err();
+        assert_eq!(
+            "Expected type: `Point`, but found `LineString`",
+            format!("{}", error)
+        )
+    }
+
+    #[test]
     fn feature_collection_with_geom_collection() {
         let geojson_str = json!({
             "type": "FeatureCollection",
@@ -697,7 +716,7 @@ mod tests {
         let geojson_polygon = Value::Polygon(geojson_multi_line_string_type1);
         let _: geo_types::Polygon<f64> = (&geojson_polygon).try_into()?;
 
-       let geojson_line_string_type1 = vec![
+        let geojson_line_string_type1 = vec![
             coord1.clone(),
             coord2.clone(),
             coord3.clone(),
@@ -714,7 +733,7 @@ mod tests {
             vec![geojson_line_string_type1],
             vec![geojson_line_string_type2],
         ]);
-        let _: geo_types::MultiPolygon<f64> =  (&geojson_multi_polygon).try_into()?;
+        let _: geo_types::MultiPolygon<f64> = (&geojson_multi_polygon).try_into()?;
 
         let geojson_geometry_collection = Value::GeometryCollection(vec![
             Geometry::new(geojson_multi_point),